重构了代码的封装和规范
